backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.1
defusedxml==0.7.0rc2
Django==3.1.5
django-cors-headers==3.7.0
django-jsonfield==1.4.1
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pathspec==0.8.1
Pillow==8.1.1
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
semantic-version==2.8.5
setuptools-rust==0.11.6
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3